Create directories for classes and function
