+++ URDF_Exporter/core/Write.py
@@ -1,0 +1,262 @@
+# -*- coding: utf-8 -*-
+"""
+Created on Sun May 12 20:46:26 2019
+
+@author: syuntoku
+"""
+
+import adsk, os
+from xml.etree.ElementTree import Element, SubElement
+from . import Link, Joint
+from ..utils import utils
+
+def write_link_urdf(joints_dict, repo, links_xyz_dict, file_name, inertial_dict):
+    """
+    Write links information into urdf "repo/file_name"
+    
+    
+    Parameters
+    ----------
+    joints_dict: dict
+        information of the each joint
+    repo: str
+        the name of the repository to save the xml file
+    links_xyz_dict: vacant dict
+        xyz information of the each link
+    file_name: str
+        urdf full path
+    inertial_dict:
+        information of the each inertial
+    
+    Note
+    ----------
+    In this function, links_xyz_dict is set for write_joint_tran_urdf.
+    The origin of the coordinate of center_of_mass is the coordinate of the link
+    """
+    with open(file_name, mode='a') as f:
+        # for base_link
+        center_of_mass = inertial_dict['base_link']['center_of_mass']
+        link = Link.Link(name='base_link', xyz=[0,0,0], 
+            center_of_mass=center_of_mass, repo=repo,
+            mass=inertial_dict['base_link']['mass'],
+            inertia_tensor=inertial_dict['base_link']['inertia'])
+        links_xyz_dict[link.name] = link.xyz
+        link.make_link_xml()
+        f.write(link.link_xml)
+
+        # others
+        for joint in joints_dict:
+            name = joints_dict[joint]['child']
+            center_of_mass = \
+                [ i-j for i, j in zip(inertial_dict[name]['center_of_mass'], joints_dict[joint]['xyz'])]
+            link = Link.Link(name=name, xyz=joints_dict[joint]['xyz'],\
+                center_of_mass=center_of_mass,\
+                repo=repo, mass=inertial_dict[name]['mass'],\
+                inertia_tensor=inertial_dict[name]['inertia'])
+            links_xyz_dict[link.name] = link.xyz            
+            link.make_link_xml()
+            f.write(link.link_xml)
+
+
+def write_joint_tran_urdf(joints_dict, repo, links_xyz_dict, file_name):
+    """
+    Write joints and transmission information into urdf "repo/file_name"
+    
+    
+    Parameters
+    ----------
+    joints_dict: dict
+        information of the each joint
+    repo: str
+        the name of the repository to save the xml file
+    links_xyz_dict: dict
+        xyz information of the each link
+    file_name: str
+        urdf full path
+    """
+    
+    with open(file_name, mode='a') as f:
+        for j in joints_dict:
+            parent = joints_dict[j]['parent']
+            child = joints_dict[j]['child']
+            joint_type = joints_dict[j]['type']
+            upper_limit = joints_dict[j]['upper_limit']
+            lower_limit = joints_dict[j]['lower_limit']
+            try:
+                xyz = [round(p-c, 6) for p, c in \
+                    zip(links_xyz_dict[parent], links_xyz_dict[child])]  # xyz = parent - child
+            except KeyError as ke:
+                app = adsk.core.Application.get()
+                ui = app.userInterface
+                ui.messageBox("There seems to be an error with the connection between\n\n%s\nand\n%s\n\nCheck \
+whether the connections\nparent=component2=%s\nchild=component1=%s\nare correct or if you need \
+to swap component1<=>component2"
+                % (parent, child, parent, child), "Error!")
+                quit()
+                
+            joint = Joint.Joint(name=j, joint_type = joint_type, xyz=xyz, \
+            axis=joints_dict[j]['axis'], parent=parent, child=child, \
+            upper_limit=upper_limit, lower_limit=lower_limit)
+            joint.make_joint_xml()
+            joint.make_transmission_xml()
+            f.write(joint.joint_xml)
+            f.write(joint.tran_xml)
+
+
+def write_gazebo_plugin_and_endtag(file_name):
+    """
+    Write about gazebo_plugin and the </robot> tag at the end of the urdf
+    
+    
+    Parameters
+    ----------
+    file_name: str
+        urdf full path
+    """
+    with open(file_name, mode='a') as f:
+        # gazebo plugin
+        gazebo = Element('gazebo')
+        plugin = SubElement(gazebo, 'plugin')
+        plugin.attrib = {'name':'control', 'filename':'libgazebo_ros_control.so'}
+        gazebo_xml = "\n".join(utils.prettify(gazebo).split("\n")[1:])
+        f.write(gazebo_xml)
+        f.write('</robot>\n')
+        
+
+def write_urdf(joints_dict, links_xyz_dict, inertial_dict, package_name, save_dir, robot_name):
+    file_name = save_dir + '/' + robot_name + '.urdf'  # the name of urdf file
+    repo = package_name + '/' + robot_name + '/bin_stl/'  # the repository of binary stl files
+    with open(file_name, mode='w') as f:
+        f.write('<?xml version="1.0" ?>\n')
+        f.write('<robot name="{}">\n'.format(robot_name))
+
+    write_link_urdf(joints_dict, repo, links_xyz_dict, file_name, inertial_dict)
+    write_joint_tran_urdf(joints_dict, repo, links_xyz_dict, file_name)
+    write_gazebo_plugin_and_endtag(file_name)
+
+
+def write_gazebo_launch(robot_name, save_dir):
+    """
+    write gazebo launch file "save_dir/launch/robot_name_gazebo.launch"
+    
+    
+    Parameter
+    ---------
+    robot_name: str
+        name of the robot
+    save_dir: str
+        path of the repository to save
+    """
+    
+    try: os.mkdir(save_dir + '/launch')
+    except: pass     
+    
+    launch = Element('launch')
+    param = SubElement(launch, 'param')
+    param.attrib = {'name':'robot_description', 'textfile':'$(find fusion2urdf)/{}/{}.urdf'.format(robot_name, robot_name)}
+
+    include_ =  SubElement(launch, 'include')
+    include_.attrib = {'file':'$(find gazebo_ros)/launch/empty_world.launch'}        
+    
+    number_of_args = 5
+    args = [None for i in range(number_of_args)]
+    args_name_value_pairs = [['paused', 'false'], ['use_sim_time', 'true'],
+                             ['gui', 'true'], ['headless', 'false'], 
+                             ['debug', 'false']]
+                             
+    for i, arg in enumerate(args):
+        arg = SubElement(include_, 'arg')
+        arg.attrib = {'name' : args_name_value_pairs[i][0] , 
+        'value' : args_name_value_pairs[i][1]}
+
+    node = SubElement(launch, 'node')
+    node.attrib = {'name':'spawn_urdf', 'pkg':'gazebo_ros', 'type':'spawn_model',\
+                    'args':'-param robot_description -urdf -model {}'.format(robot_name)}
+    
+    launch_xml = "\n".join(utils.prettify(launch).split("\n")[1:])        
+    
+    file_name = save_dir + '/launch/' + robot_name + '_gazebo.launch'    
+    with open(file_name, mode='w') as f:
+        f.write(launch_xml)
+
+
+def write_control_launch(robot_name, save_dir, joints_dict):
+    """
+    write control launch file "save_dir/launch/robot_name_control.launch"
+    
+    
+    Parameter
+    ---------
+    robot_name: str
+        name of the robot
+    save_dir: str
+        path of the repository to save
+    joints_dict: dict
+        information of the joints
+    """
+    
+    try: os.mkdir(save_dir + '/launch')
+    except: pass     
+    
+    launch = Element('launch')
+
+    controller_name = robot_name + '_controller'
+    rosparam = SubElement(launch, 'rosparam')
+    rosparam.attrib = {'file':'$(find fusion2urdf)/launch/' + controller_name + '.yaml',
+                       'command':'load'}
+                       
+    controller_args_str = ' '.join([joint + '_position_controller' for joint in joints_dict]) \
+                            + ' joint_state_controller'
+    node_controller = SubElement(launch, 'node')
+    node_controller.attrib = {'name':'controller_spawner', 'pkg':'controller_manager', 'type':'spawner',\
+                    'respawn':'false', 'output':'screen', 'ns':robot_name,\
+                    'args':'{}'.format(controller_args_str)}
+    
+    node_publisher = SubElement(launch, 'node')
+    node_publisher.attrib = {'name':'robot_state_publisher', 'pkg':'robot_state_publisher',\
+                    'type':'robot_state_publisher', 'respawn':'false', 'output':'screen'}
+    remap = SubElement(node_publisher, 'remap')
+    remap.attrib = {'from':'/joint_states',\
+                    'to':'/' + robot_name + '/joint_states'}
+    
+    launch_xml = "\n".join(utils.prettify(launch).split("\n")[1:])        
+    
+    file_name = save_dir + '/launch/' + robot_name + '_control.launch'    
+    with open(file_name, mode='w') as f:
+        f.write(launch_xml)
+        
+
+def write_yaml(robot_name, save_dir, joints_dict):
+    """
+    write yaml file "save_dir/launch/robot_name_controller.yaml"
+    
+    
+    Parameter
+    ---------
+    robot_name: str
+        name of the robot
+    save_dir: str
+        path of the repository to save
+    joints_dict: dict
+        information of the joints
+    """
+    try: os.mkdir(save_dir + '/launch')
+    except: pass 
+
+    controller_name = robot_name + '_controller'
+    file_name = save_dir + '/launch/' + controller_name + '.yaml'
+    with open(file_name, 'w') as f:
+        f.write(controller_name + ':\n')
+        # joint_state_controller
+        f.write('  # Publish all joint states -----------------------------------\n')
+        f.write('  joint_state_controller:\n')
+        f.write('    type: joint_state_controller/JointStateController\n')  
+        f.write('    publish_rate: 50\n\n')
+        # position_controllers
+        f.write('  # Position Controllers --------------------------------------\n')
+        for joint in joints_dict:
+            f.write('    ' + joint + '_position_controller:\n')
+            f.write('      type: effort_controllers/JointPositionController\n')
+            f.write('      joint: '+ joint + '\n')
+            f.write('      pid: {p: 100.0, i: 0.01, d: 10.0}\n')
+
